fix: [sc-11] move logout functionality from user to auth route
