+++ src/routes/users.route.js
@@ -3,7 +3,6 @@ const express = require('express');
 const {
   getUserProfile,
   updateUserProfile,
-  logout,
 } = require('../controllers/users.controller.js');
 
 const { verifyToken } = require('../middlewares/auth.middleware.js');
@@ -13,7 +12,5 @@ const router = express.Router();
 router.get('/me', verifyToken, getUserProfile);
 
 router.put('/me', verifyToken, updateUserProfile);
-
-router.post('/logout', verifyToken, logout);
 
 module.exports = router;
